feat: replace ArticleItemCard with ArticleCard
